Simplify calls to findOverlaps() by not explicitly setting optional arguments to their default values.
Default values for 'maxgap' and 'minoverlap' arguments have changed in the
IRanges package for findOverlaps() and family, but the default behavior of
findOverlaps() and family remains the same (except on zero-width ranges).
As a consequence of this change, explicitly passing the old 'maxgap' and
'minoverlap' defaults in a call to findOverlaps() now produces an error.
NOT explicitly setting these optional arguments to their default values
addresses the issue.